+++ NEWS.md
@@ -137,6 +137,7 @@ Red Panda C++ Version 2.27
   - enhancement: Auto hide Edit/Selection/Code/Refactor menu if no file openning.
   - enhancement: Auto hide Project menu if no project openning.
   - fix: Toggle breakpoint by shortcut may use wrong line.
+  - fix: Size of the icons in problem and problem set panel are not correct.
     
 Red Panda C++ Version 2.26
   - enhancement: Code suggestion for embedded std::vectors.